==> Part 1_Weather_Database.ipynb <==
import pandas as pd
import numpy as np
import json
import requests
import random 
import time
import gmaps
import matplotlib.pyplot as plt 
from citipy import citipy
from config import weather_api_key
from datetime import datetime

lats = np.random.uniform(low=-90.000, high=90.000, size=2000)
lngs = np.random.uniform(low=-180.000, high=180.000, size=2000)
lat_lngs = zip(lats, lngs)
coordinates = list(lat_lngs)


# Create a list for holding the cities.
cities = []
# Identify the nearest city for each latitude and longitude combination.
for coordinate in coordinates:
    city = citipy.nearest_city(coordinate[0], coordinate[1]).city_name

    # If the city is unique, then we will add it to the cities list.
    if city not in cities:
        cities.append(city)
# Print the city count to confirm sufficient count.
len(cities)

# Import the datetime module from the datetime library.
from datetime import datetime
# Get the date from the JSON file.
date = boston_data["dt"]
# Convert the UTC date to a date format with year, month, day, hours, minutes, and seconds.
datetime.utcfromtimestamp(date)
datetime.utcfromtimestamp(date).strftime('%Y-%m-%d %H:%M:%S')

city_data = {"City": [] , "Lat": [] , "Lng": [] , "Max Temp": [] , "Humidity": [] , "Cloudiness": [] , "Wind Speed": [] , "Country": [] , "Date": []}

for i in range(len(cities)):
    # Create endpoint URL with each city.
    city_name = cities[i]
    url = f'https://api.openweathermap.org/data/2.5/weather?q={city_name}&appid={weather_api_key}'
    time.sleep(1.25)
    try: 
        city_weather = requests.get(url).json()
        # Parse out the needed data.
        name = city_weather["name"]
        city_data["City"].append(name)

        city_lat = city_weather["coord"]["lat"]
        city_data["Lat"].append(city_lat)

        city_lng = city_weather["coord"]["lon"]
        city_data["Lng"].append(city_lng)

        city_max_temp_kel = city_weather["main"]["temp_max"]
        city_max_temp = city_max_temp_kel - 273.15
        city_data["Max Temp"].append(city_max_temp)

        city_humidity = city_weather["main"]["humidity"]
        city_data["Humidity"].append(city_humidity)

        city_clouds = city_weather["clouds"]["all"]
        city_data["Cloudiness"].append(city_clouds)

        city_wind = city_weather["wind"]["speed"]
        city_data["Wind Speed"].append(city_wind)

        city_country = city_weather["sys"]["country"]
        city_data["Country"].append(city_country)

            # Convert the date to ISO standard.
        city_date = datetime.utcfromtimestamp(city_weather["dt"]).strftime('%Y-%m-%d %H:%M:%S')
        city_data["Date"].append(city_date)
    except:
        print(f"City '{city_name}' NOT found on OpenWeatherAPI cloud.")
        pass

city_data_df = pd.DataFrame(city_data)
city_data_df.head()



# Create the output file (CSV).
output_data_file = "weather_data/WeatherPy_database.csv"
# Export the City_Data into a CSV.
city_data_df.to_csv(output_data_file, index_label="City_ID")
