==> Part 2_Travel_Destination_Map.ipynb <==
## Deliverable 2. Create a Customer Travel Destinations Map.

# Dependencies and Setup
import pandas as pd
import requests
import gmaps
import numpy as np
import matplotlib.pyplot as plt
# Import API key
from config import g_key

# Configure gmaps API key
gmaps.configure(api_key=g_key)

# 1. Import the WeatherPy_database.csv file. 
city_data_df = pd.read_csv("weather_data/WeatherPy_database.csv")
city_data_df.head()

# 2. Prompt the user to enter minimum and maximum temperature criteria 
print("Enter the maximum preffered temperature for your travel destination:")
max_t = input()

print("Enter the minimum preffered temperature for your travel destination:")
min_t = input()



max_t = 32
min_t = 20

# 3. Filter the city_data_df DataFrame using the input statements to create a new DataFrame using the loc method.
preferred_cities_df = city_data_df.loc[(city_data_df["Max Temp"] <= max_t) & \
                                       (city_data_df["Max Temp"] >= min_t)]
preferred_cities_df.head(10)

# 4a. Determine if there are any empty rows.
[preferred_cities_df.iloc[i,j] for i,j in zip(*np.where(pd.isnull(preferred_cities_df)))]

# 4b. Drop any empty rows and create a new DataFrame that doesn’t have empty rows.
preferred_cities_df = preferred_cities_df.dropna()
preferred_cities_df.head()

# 5a. Create DataFrame called hotel_df to store hotel names along with city, country, max temp, and coordinates.
hotel_df = preferred_cities_df[["City", "Country", "Max Temp", "Lat", "Lng"]].copy()

# 5b. Create a new column "Hotel Name"
hotel_df["Hotel Name"] = ""
hotel_df.head()

# 6a. Set parameters to search for hotels with 5000 meters.
from config import p_key
params = {
    "radius": 5000,
    "types": "lodging",
    "key": p_key,
    "location": ''}
    
# Iterate through the DataFrame.
for index, row in hotel_df.iterrows():
    # Get the latitude and longitude.
    lat = row["Lat"]
    lng = row["Lng"]

    # Add the latitude and longitude to location key for the params dictionary.
    params["location"] = f"{lat},{lng}"

    # Use the search term: "lodging" and our latitude and longitude.
    base_url = "https://maps.googleapis.com/maps/api/place/nearbysearch/json"
    # Make request and get the JSON data from the search.
    hotels = requests.get(base_url, params=params).json()
    # Grab the first hotel from the results and store the name.
    # Grab the first hotel from the results and store the name.
    try:
        hotel_df.loc[index, "Hotel Name"] = hotels["results"][0]["name"]
    except (IndexError):
        print("Hotel not found... skipping.")

hotel_df.head()

        

hotel_df

# 7. Drop the rows where there is no Hotel Name.
missing_hotels_df = hotel_df.loc[(hotel_df['Hotel Name'] == '')]
voidlist = missing_hotels_df.index
hotel_df = hotel_df.drop(voidlist)
#Conclusion: No empty hotel name in the dataframe

# 8a. Create the output File (CSV)
output_data_file = "vaccation_data/WeatherPy_vacation.csv"
# 8b. Export the City_Data into a csv
hotel_df.to_csv(output_data_file, index_label="City_ID")

# 9. Using the template add city name, the country code, the weather description and maximum temperature for the city.
info_box_template = """
<dl>
<dt>Hotel Name</dt><dd>{Hotel Name}</dd>
<dt>City</dt><dd>{City}</dd>
<dt>Country</dt><dd>{Country}</dd>
<dt>Max Temp</dt><dd>{Max Temp} °C</dd>
</dl>
"""

# Store the DataFrame Row.
hotel_info = [info_box_template.format(**row) for index, row in hotel_df.iterrows()]

# Add a heatmap of temperature for the vacation spots and a pop-up marker for each city.
locations = hotel_df[["Lat", "Lng"]]
max_temp = hotel_df["Max Temp"]
fig = gmaps.figure(center=(30.0, 31.0), zoom_level=1.5)
heat_layer = gmaps.heatmap_layer(locations, weights=max_temp,dissipating=False,
             max_intensity=300, point_radius=4)
marker_layer = gmaps.marker_layer(locations, info_box_content=hotel_info)
fig.add_layer(heat_layer)
fig.add_layer(marker_layer)
fig
