==> Part 3_Travel_Itinerary_Map.ipynb <==
## Deliverable 3. Create a Travel Itinerary Map.

# Dependencies and Setup
import pandas as pd
import requests
import gmaps
import numpy as np
# Import API key
from config import g_key

# Configure gmaps
gmaps.configure(api_key=g_key)

# 1. Read the WeatherPy_vacation.csv into a DataFrame.
vacation_df = pd.read_csv("vaccation_data/WeatherPy_vacation.csv")
vacation_df.head()

# 2. Using the template add the city name, the country code, the weather description and maximum temperature for the city.
info_box_template = """
<dl>
<dt>Hotel Name</dt><dd>{Hotel Name}</dd>
<dt>City</dt><dd>{City}</dd>
<dt>Country</dt><dd>{Country}</dd>
<dt>Max Temp</dt><dd>{Max Temp} °C</dd>
</dl>
"""

# 3a. Get the data from each row and add it to the formatting template and store the data in a list.
hotel_info = [info_box_template.format(**row) for index, row in vacation_df.iterrows()]

# 3b. Get the latitude and longitude from each row and store in a new DataFrame.
locations = vacation_df[["Lat", "Lng"]]
max_temp = vacation_df["Max Temp"]

# 4a. Add a marker layer for each city to the map.
fig = gmaps.figure(center=(30.0, 31.0), zoom_level=1.5)
heat_layer = gmaps.heatmap_layer(locations, weights=max_temp,dissipating=False,
             max_intensity=300, point_radius=4)
marker_layer = gmaps.marker_layer(locations, info_box_content=hotel_info)
fig.add_layer(heat_layer)
fig.add_layer(marker_layer)

# 4b. Display the figure
fig


# From the map above pick 4 cities and create a vacation itinerary route to travel between the four cities. 
# 5. Create DataFrames for each city by filtering the 'vacation_df' using the loc method. 
# Hint: The starting and ending city should be the same city.

vacation_start = vacation_df.loc[(vacation_df["City"] == 'Jasper') & (vacation_df["Country"] == 'US')]
vacation_end = vacation_df.loc[(vacation_df["City"] == 'Jasper') & (vacation_df["Country"] == 'US')]
vacation_stop1 = vacation_df.loc[(vacation_df["City"] == 'Hamilton') & (vacation_df["Country"] == 'US')]
vacation_stop2 = vacation_df.loc[(vacation_df["City"] == 'Marshall') & (vacation_df["Country"] == 'US')] 
vacation_stop3 = vacation_df.loc[(vacation_df["City"] == 'Estelle') & (vacation_df["Country"] == 'US')] 


# 6. Get the latitude-longitude pairs as tuples from each city DataFrame.
vacation_start_lat = float(vacation_start['Lat'])
vacation_start_lng = float(vacation_start['Lng'])
vacation_start_cor = (vacation_start_lat , vacation_start_lng)

vacation_end_lat = float(vacation_end['Lat'])
vacation_end_lng = float(vacation_end['Lng'])
vacation_end_cor = (vacation_end_lat , vacation_end_lng)

vacation_stop1_lat = float(vacation_stop1['Lat'])
vacation_stop1_lng = float(vacation_stop1['Lng'])
vacation_stop1_cor = (vacation_stop1_lat , vacation_stop1_lng)

vacation_stop2_lat = float(vacation_stop2['Lat'])
vacation_stop2_lng = float(vacation_stop2['Lng'])
vacation_stop2_cor = (vacation_stop2_lat , vacation_stop2_lng)

vacation_stop3_lat = float(vacation_stop3['Lat'])
vacation_stop3_lng = float(vacation_stop3['Lng'])
vacation_stop3_cor = (vacation_stop3_lat , vacation_stop3_lng)

print(f'_________________________________________________ \n***Vaccation starting/ending point coordinates: \n {vacation_start_cor} \n_________________________________________________  \n***Stop over coordinates: \n {vacation_stop1_cor} \n {vacation_stop2_cor} \n {vacation_stop3_cor}')

# 7. Create a direction layer map using the start and end latitude-longitude pairs,
# and stop1, stop2, and stop3 as the waypoints. The travel_mode should be "DRIVING", "BICYCLING", or "WALKING".
from config4 import dir_key

fig_driving = gmaps.figure()
driving_directions = gmaps.directions_layer(vacation_start_cor, vacation_end_cor, waypoints= [vacation_stop1_cor, vacation_stop2_cor, vacation_stop3_cor] , travel_mode= 'DRIVING')

fig_driving.add_layer(driving_directions)
fig_driving


# 8. To create a marker layer map between the four cities.
#  Combine the four city DataFrames into one DataFrame using the concat() function.
itinerary_df = pd.concat([vacation_start, vacation_end, vacation_stop1, vacation_stop2, vacation_stop3],ignore_index=True)
itinerary_df

# 9 Using the template add city name, the country code, the weather description and maximum temperature for the city. 
info_box_template = """
<dl>
<dt>Hotel Name</dt><dd>{Hotel Name}</dd>
<dt>City</dt><dd>{City}</dd>
<dt>Country</dt><dd>{Country}</dd>
<dt>Max Temp</dt><dd>{Max Temp} °C</dd>
</dl>
"""

# 10a Get the data from each row and add it to the formatting template and store the data in a list.
hotel_info = [info_box_template.format(**row) for index, row in itinerary_df.iterrows()]

# 10b. Get the latitude and longitude from each row and store in a new DataFrame.
locations = itinerary_df[["Lat", "Lng"]]
max_temp = itinerary_df["Max Temp"]
heat_layer = gmaps.heatmap_layer(locations, weights=max_temp,dissipating=False,
             max_intensity=300, point_radius=4)
#add a marker layer
marker_layer = gmaps.marker_layer(locations, info_box_content=hotel_info)
fig_driving.add_layer(heat_layer)
fig_driving.add_layer(marker_layer)
fig_driving
